add import option risk data (by csv) modify importexcle

+++ importEntrust.py
@@ -7,26 +7,49 @@ import os
 import hq as hq
 
 FORMAT_date = lambda x: x.replace('-','')
+FORMAT_SPACE = lambda x: x.strip()
 DATE = time.strftime('%Y%m%d',time.localtime())
 DATE = DATE[:4]+'年'+DATE[4:6]+'月'+DATE[6:8]+'日'
-csv_path =u'C:\\'+DATE+'.csv'
-excel_path = u'C:\\'+DATE+'.xlsx'
+
 #DATE = '180812'
 #print(DATE)
 #path = "C:/image/" + DATE + "/"
 
 def csv_to_excle():
-    csv = pd.read_csv(csv_path, encoding='gb18030')
-    csv.to_excel(excel_path, sheet_name=DATE)
+    if (os.path.exists(csv_path)):
+        csv = pd.read_csv(csv_path, encoding='gb18030')
+        csv.to_excel(excel_path, sheet_name=DATE)
+    else:
+        csv = pd.read_csv("c:\\risk.csv", encoding='gb18030')
+        csv.to_excel("c:\\risk.xlsx", sheet_name='risk')
 
-def importexcel(date):
+def importentrust(date):
+    csv_path = u'C:\\' + DATE + '.csv'
+    excel_path = u'C:\\' + DATE + '.xlsx'
     sql = "delete from t_entrust where date = '"+date+"'"
     hq._excutesql(sql)
     #df = pd.read_csv(excel_path, index_col=False, quoting=3, sep=" ",encoding='gb18030')
-    csv_to_excle()
+    #csv_to_excle()
+    pd.read_csv(csv_path, encoding='gb18030').to_excel(excel_path, sheet_name=DATE) #csv转excle
     df = pd.read_excel(excel_path,sheet_name=DATE)
     df.drop(df.columns[len(df.columns) - 1], axis=1, inplace=True)
     df.columns = ['date', 'time', 'market', 'code', 'name', 'deal', 'type', 'price', 'quantity', 'amount','reserve_flag', 'serial', 'remark','fee']
     df['date'] = df['date'].map(FORMAT_date)
     df.to_sql('t_entrust', c.ENGINE, if_exists='append')
     os.remove(excel_path)
+    os.remove(csv_path)
+
+def importrisk(date):
+    csv_path = u'C:\\risk.csv'
+    sql = "delete from t_opt_risk where date = '" + date + "'"
+    hq._excutesql(sql)
+    df = pd.read_csv(csv_path,encoding='gb18030')
+    df.drop(df.columns[len(df.columns) - 1], axis=1, inplace=True)
+    df.reset_index(level=0, inplace=True)
+    df['index'] =df['index'].map(FORMAT_date)
+    df.columns = ['date', 'ts_code', 'opt_code', 'name', 'Delta', 'Theta', 'Gamma', 'Vega', 'Rho']
+    df['date'] = df['date'].map(FORMAT_SPACE)
+    df['opt_code'] = df['opt_code'].map(FORMAT_SPACE)
+    df['name'] = df['name'].map(FORMAT_SPACE)
+    df.to_sql('t_opt_risk', c.ENGINE, if_exists='append')
+    os.remove(csv_path)
